refactoring and new registry integration
