+++ jaclang/plugin/default.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 
 import fnmatch
-import json
 import os
+import pickle
 import types
 from dataclasses import field
 from functools import wraps
@@ -12,7 +12,15 @@ from typing import Any, Callable, Optional, Type
 
 from jaclang.compiler.absyntree import Module
 from jaclang.compiler.constant import EdgeDir, colors
-from jaclang.core.aott import aott_lower, aott_raise
+from jaclang.core.aott import (
+    aott_lower,
+    aott_raise,
+    extract_non_primary_type,
+    get_all_type_explanations,
+    get_info_types,
+    get_object_string,
+    get_type_annotation,
+)
 from jaclang.core.construct import (
     Architype,
     DSFunc,
@@ -29,14 +37,8 @@ from jaclang.core.construct import (
     root,
 )
 from jaclang.core.importer import jac_importer
-from jaclang.core.utils import (
-    extract_non_primary_type,
-    filter,
-    get_all_type_explanations,
-    get_object_string,
-    get_type_annotation,
-    traverse_graph,
-)
+from jaclang.core.registry import Scope
+from jaclang.core.utils import traverse_graph
 from jaclang.plugin.feature import JacFeature as Jac
 from jaclang.plugin.spec import T
 
@@ -439,8 +441,8 @@ class JacFeatureDefaults:
         model: Any,  # noqa: ANN401
         model_params: dict[str, Any],
         scope: str,
-        incl_info: tuple[str, str],
-        excl_info: tuple,
+        incl_info: list[tuple[str, str]],
+        excl_info: list[tuple[str, str]],
         inputs: tuple,
         outputs: tuple,
         action: str,
@@ -450,18 +452,18 @@ class JacFeatureDefaults:
             os.path.join(
                 os.path.dirname(file_loc),
                 "__jac_gen__",
-                os.path.basename(file_loc).replace(".jac", "_registry.json"),
+                os.path.basename(file_loc).replace(".jac", ".registry.pkl"),
             ),
-            "r",
+            "rb",
         ) as f:
-            registry_data = json.load(f)
-
+            mod_registry = pickle.load(f)
+        _scope = Scope.get_scope_from_str(scope)
         reason = False
         if "reason" in model_params:
             reason = model_params.pop("reason")
 
         type_collector: list = []
-        information, collected_types = filter(scope, registry_data, incl_info)
+        information, collected_types = get_info_types(_scope, mod_registry, incl_info)
         type_collector.extend(collected_types)
 
         inputs_information_list = []
@@ -477,7 +479,7 @@ class JacFeatureDefaults:
         type_collector.extend(extract_non_primary_type(outputs[2]))
 
         type_explanations_list = list(
-            get_all_type_explanations(type_collector, registry_data).values()
+            get_all_type_explanations(type_collector, mod_registry).values()
         )
         type_explanations = "\n".join(type_explanations_list)
